Update patch set 2
Patch Set 2: Code-Review-1

Correct me if I'm wrong, but this assumes that the first script in the page that has a nonce="" is the dev_mode_on.js, which cannot be guaranteed, for various reasons:

* frontpage.html's bookmarklet script injects dev_mode_js at the *end* of the <head>
* it injects it without a nonce=""

If anything, this change should also update frontpage.html in some way to fix those two points. But even then, I'm not sure it'll work in every case.

BTW, if possible, I think we should keep the Compile bookmarklet (which would be blocked by CSP, but could be still added as a bookmark I suppose) and handle click events to replace the action with a CSP-compatible one.

Patch-set: 2
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=-1