Update patch set 2
Patch Set 2:

I am not sure I get your concern. In my view, this would be loaded in the debug version of a module, and my added code would just get the first non-null nonce of a <script> element. I am not a GWT expert, but the team I worked with tested this and was happy.

I have no objections in keeping it a bookmarklet and just adding the onclick as a CSP-compatible handler - in fact, this is what I was doing before :) .

Patch-set: 2
Reviewer: Gerrit User 10090 <10090@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0