Update patch set 2
Patch Set 2:

If we're counting bytes, I guess I'll be surprised to see them using spliterators and streams to begin with (and as far as I'm aware, these are the only APIs that use this new convention of AIOOBE for start > end). Doubly so if we're looking at perf numbers - streams are slower on standard jvm, and i'd expect it to surprise no one if they are slower in JS as well. 

As long as we're still pruning methods that aren't used, only projects that decide to already have paid this cost anyway will see anything - if (a < b) {throw c("string");} - the c type won't be new (other throws already must use it), the string could be the same as what we already for the other checks. Factor out the other checks to their own method and have both variants call them?

Otherwise we might as well just make everything throw IllegalStateException (and maaaybe a few NPEs) if we think the actual error type doesn't really matter. But we don't, and I assume this is for a reason.

Patch-set: 2
Label: Verified=0