Update patch set 5
Patch Set 5:

Ping.

Making the methods JsVarags or not depending on canBeReferencedExternally potentially introduces inconsistencies between overrides.

In the end I decided not to tighten the parameter if it is JsVarargs which is a correct semantic and is consistent with the behaviour that for JsVarargs there is an implicit  array creation.

Patch-set: 5