Make each method reference be its own different lambda.
Even method references to same method (with or without explicit
qualifier) might need different impelementations due to generics.

Instead of attempting to reuse, this patch takes the simples approach
and considers each method reference a different implementation in
the same way lambdas are treated.

A future implementation could be introduced to deduplicate lambdas
and that would not only optimize member references.

Change-Id: Ifef2bb703ad4172977e965f786a0807bdccd5f7b