Fix runtime type error due to missing lambda implementation method.
Due to the current startegy of implementing the code for the lambda
method in the enclosing class, if the lambda refers to the enclosing
"this" and is declared in a default method in an interface, it ends
up as a private instance method in such interface.

Interfaces are not expected to have any instance methods (appart from
default methdods) and thus the implementation method was never
emitted.

This is yet another hack which makes those methods devirtualized.
The better solution for this and other bugs related to lambdas is
to syntehsize these methods in the inner class created for the
lambda instead.

Change-Id: Ie59bb5df3c3b5ba29f3616f75abb1df1d0f7366c