Fix incorrect pruning of overridden JsMethod.
JsMember information is inherited from overriden methods
and pruning any of these overridden methods might cause the
compiler to "forget" that a JsMethod/JsProperty is a
JsMethod/JsProperty.

This is the correct behaviour in most cases (as only if
-nogenerateJsInteropExports those overriden members are pruned),
except for the case of native JsMethods.

This patch makes overriden native JsMembers reachable if the
overriding JsMember is deemed reachable.

Bug: #9358
Bug-Link: http://github.com/gwtproject/gwt/issues/9358
Change-Id: I80637883849af5d99542f8d0bc35c3e5b895a9e1