Fix packaging and Maven dependencies after sourcemaps update
Closure Compiler (hence sourcemaps) has switched to GSON for JSON
processing, and GSON was, as a result, bundled into gwt-dev and
requestfactory-server in Maven JARs.
This change removes it entirely from requestfactory-server, adds it to
gwt-servlet-deps instead, and unbundles it from gwt-dev when publishing
to Maven.

Also, Colt was erroneously added to gwt-servlet.jar (not even a
server-side dependency).

Change-Id: I34b55c0dc26dd2d960cb24e530f5acd8b88c0032