Update patch set 2
Patch Set 2:

Jens, I agree that it would have been ideal to have the regex support in generateJsInteropExports before this.
However, as Ray stated, this is not too much bloat since it will only keep a few extra methods in collections hanging assuming they already using these APIs and the types are not pruned.

Patch-set: 2