Update patch set 1
Patch Set 1:

> I obviously support getting the build back to green, but that
 > sounds like "legacy" DevMode is currently completely broken right?
 > Or is this only one (or a few) specific part that is broken but
 > causes many other tests to fail?

The issue here is that the support for devirtualized methods in JSNI (https://gwt-review.googlesource.com/#/c/15121) is not implemented in DevMode. Classes that contain JSNI that makes use of this will fail to load in DevMode; in particular the test for this feature in JsniDispatchTest. There is a provision for preventing running certain tests in devmode with @DoNotRun but not to prevent loading.

Because devmode is deprecated I think it is OK to disable the Suites containing tests failing for this reason.

Other possible approaches (if someone wants to invest in them) would either to create a Suite with all the tests that break devmode or to fix ComilingClassLoader/  JsSourceGenWithJsniIdentFixup to error on calls to these methods but not loading.

Patch-set: 1
Reviewer: Gerrit User 5105 <5105@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0