Update patch set 2
Patch Set 2:

> If instead you are suggesting that consensus is to just forbid
 > access to default and static methods and pretend they don't exist,
 > I'd have to disagree at least on the grounds that this makes
 > UiBinder/GssResource unpredictable at best (interfaces extending
 > interfaces, etc), or at worst broken and unmaintained. Unless
 > someone here is jumping on the work of porting either/both of these
 > to APT...?

I have to disagree, I still think that hiding them is the safest option. Generators don't expect to see default and static methods in interfaces and some assume that all the methods found in interfaces are to be implemented.

That is definitely broken behaviour for interfaces that declare static methods (we have already seem breakages due to this in a mock generator). So clearly there should be no objection to omit static interface methods.

W.r.t to default methods, because they are intended for adding functionality to interfaces without needed to change implementors, in most cases they can be safely ignored as they don't need to be implemented.

Patch-set: 2