Update patch set 2
Patch Set 2:

No disagreement that existing code may misbehave - but it won't be the first time that we've pushed things that weren't 100% out of the box in an updated version adding new interface methods or final methods to classes, or the rebuild that CssResource received. We're not averse to breaking things when it means dramatically improving them.

The only argument I have is that things that reasonably _should_ work now won't: gssresource making a static interface call, uibinder's {} notation trying to invoke a default method, etc. This is a surprising departure - we refuse to deprecate Generators, and selectively add new features and refuse others.

We claim that we have publicly declared "Do not use generators anymore, use APTs.
" for a year and a half, except when it suits us. At least one notable generator has still gotten periodic maintenance - GssResource has had several fixes and even new features made to it over the past year instead of the correct option of porting it. 

I'm expecting that as the only organization within GWT pushing back about this, that we'll see more transparency around features, and around so-called deprecated tools. I'm disappointed that we still have yet to see a single annotation processor (or even a way to kick off that process from within super dev mode) from Google despite all of the noise about how incredible they are and how superior they are to generators. Thomas ported places, I ported SafeHtmlTemplates, I'm assisting another contributor to bring Editors out right now, and RedHat has expressed an interest in validation. 

"Safe" is a valid position to take, but its frustrating to see it only applied intermittently. I don't disagree with the decision that has been made, just with how we've had to "discuss" it, moving to a public forum, and then deciding that technical matters here in the review override all.

Patch-set: 2