Update patch set 2
Patch Set 2:

> "Safe" is a valid position to take, but its frustrating to see it
 > only applied intermittently. I don't disagree with the decision
 > that has been made, just with how we've had to "discuss" it, moving
 > to a public forum, and then deciding that technical matters here in
 > the review override all.

Not to flog on a dead horse :) We are in the process of trying to close the release and this issue needed a prompt, practical and safe solution.

Sorry for not giving pointers but my familiarity with the typeoracle/generator interface is extremely limited. I still think that generators that don't explicitly ask for Java 8 features should get a safe Java 7 view of the classes (much in the same vein ASM does it). 

The way I would approach providing a Java 8 view is having GeneratorContext.getTypeOracle(version) and making JRealClassType a proxy over JRealClassType8 or something like that. Having a parallel structure so that once you one of those classes they are just proxies over the Java 8 ones filtering. I'd gladly review it.

That being said I think it is quite some effort and probably a stretch to get it implemented and well tested in the 2.8 timeframe.

Patch-set: 2