Update patch set 2
Patch Set 2:

Complete agreement with you there - its better to get this release out the door (and to have my viewpoint lose out for the sake of expediency). I'd love to see a `asJava8()`-like method that removes these filters, but recognize that given the choice between holding back the release and limiting an API that should be on its way out anyway, there isn't much of a choice.

Patch-set: 2