Update patch set 2
Patch Set 2:

> I'm wondering why not the getAll is the default behavior since it
 > would fix other broken cases like in AutBean (that I guess could be
 > many), and the other way around would be used for specific cases
 > wanting the old (legacy?) API.

Concern was backwards compatibility: more interfaces have gained new static and default methods than have "lost" mandatory-to-implement methods that became default ones, and the risk that generators worked with the assumption that all methods in an interface were mandatory to implement is far from non-zero. The risk is actually probably more important for static methods, where a generator would generate an instance method with an @Override which would break the compilation. The theory was that current code should continue to work as-is and Java 8 additions might not be supported at all (hence: hidden to generators), and generators should move to annotation processors if they want to take advantage of the new language features; in practice, it's not that straightforward (but generating an Iterator implementation is kind of an edge-case; I was actually surprised AutoBean did it). Also, AutoBean is also an edge-case in that it generates proxies to other classes, where I believe most people would prefer default methods to be ignored so they could add "utility methods" to their interfaces (e.g. add a static or default method to a Constants or Messages or ClientBundle interface itself instead of using a companion helper class)

See discussion at https://groups.google.com/d/msg/google-web-toolkit-contributors/E_jINcSyd4o/JXVsOBB0BAAJ

Patch-set: 2
Label: Verified=0