Update patch set 3
Patch Set 3:

I had a discussion with rluble@ on the idea of exposing java8 features to generators and we generally do not think (even with the autobeans regression) that we should do it. Let's see if RC1 reveals problems that we did not anticipate and then revisit this. WDYT?

Patch-set: 3
Reviewer: Gerrit User 5385 <5385@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0