DeadCodeElimination was not optimizing some expressions fully.
DeadCodeElimination was missing optimizaiton opportunities because
it did not account correctly for the methods it had to be run on.

In particular when a constructor body gets optimized away there are new
optimization opportunities in places where it is invoked.

Now DeadCodeElimination is run not only on modified methods but also
on their callers.

Change-Id: I1abbee03ce420ae5a29f2bb6fa829c8eb94dff87