Implement abstract method bridges for functional expressions.
The synthetic inner class that implements lambdas and method references
might need bridges for the single abstract method (and also for
default methods).

This patch creates the bridge methods for the single abstract method,
leaving the task of creating bridges for the default methods for a
follow up patch.

Bug: #9406
Bug-Link: https://github.com/gwtproject/gwt/issues/9406
Change-Id: I80bb32b3f81e08a1f583ed78f1765df8cdd21caf