Update patch set 7
Patch Set 7:

PTAL. While the changes in patchset 3 fix the problem, I got to understand the real issue, create a test and do a better fix.

The real issue is that when Double is a reference only type (after the first compile), its $create methods are not flowed into, hence their parameters are not resolved. This is due to UnifyAst not seeing a direct reference to $create but only to the constructor Double(...) which is flowed into. (if -generateJsInteropExports was provided, then everything worked fine as all the $create methods are static JsMethod and hence traversed as entry points.)

I created a repro case and added the right behavior to UnifyAST.

Patch-set: 7
Label: Verified=0