Update patch set 5
Patch Set 5:

Wondering if GWT should make it documented by introducing a new constant in JsPackage. If you want to inject external JS into the GWT iframe, so its local to your GWT app, then you can not use it with JsInterop unless you do some namespace trickery. With <window> this would be easily possible now.

Whats the reasoning that JsInterop forces external JS to be in top/global scope? Shouldn't the developer decide it and live with the consequences?

Patch-set: 5
Label: Verified=0