Update patch set 2
Patch Set 2:

> Shouldn't you have tests for the client side? or that error does
 > not happen in browser?

This bug originally occurred in the browser.
But because it exactly behaves the same on the server as in the browser, 
I was quite pleased to find a way to unit test it without having to write a gwt test case.
Because of the long execution time of gwt test cases, I try to avoid it whenever possible. 

I'll write a gwt test case if you like, but it would be good to know what you expect to behave differently so I can actually include it in the test.

Patch-set: 2
Label: Verified=0