Update patch set 2
Patch Set 2:

Based on the code, I believe that the issue is the same in browser than in JVM.
But since client Date.java rely on the JS implementation, it might happen some unknown problem.

If you find a way to easily modify the test class so as it runs in client and server, it would be nice to have.

Otherwise, there are already tests for the format method that run in client, you could add some tests there.

Patch-set: 2