Fix loading bug in waitForBodyLoaded.js.
Page loading normally happens in two phases: first DOMContentLoaded is fired, and readyState goes to interactive, second the normal document/window.load is fired and readyState goes to complete.

isBodyLoaded.js verifies loading state by checking if readyState is complete. However, waitForBodyLoaded.js listens for DOMContentLoaded. So when that event is fired, the ready state is not there yet, thus this event handler is useless.

(GWT has a fallback mechanism, a 50ms interval timer. That works, but it causes an unnecessary latency. Also, I'm working on a project where raw JS scripts are loaded next to GWT and they assume GWT is loaded when window.load is fired. But because of this bug, this doesn't work, causing hacky workarounds.)

Repro here: https://jsfiddle.net/wxawpadr/1/ . This is mostly a copy of GWT's scripts and logs what's happening to the console. This shows that the callback is triggered by the timer, not the event.

The fix is really simple: if isBodyLoaded.js wants a complete readyState, then we listen for the readystatechange event. Also there's a minor code cleanup relating to that, because we don't want to fire when readyState goes to interactive.

Change-Id: I7c9f1c89bd12916549581ffe8bbc522c09f56b50
Review-Link: https://gwt-review.googlesource.com/#/c/16960/