Update patch set 3
Patch Set 3: Code-Review+1

We are cutting GWT 2.8 right now, so have master closed off for any non critial changes. I think this is good to go. If this is urgent you can ping me internally and we will figure out how to land this internally (without landing it here).

Patch-set: 3