Update patch set 3
Patch Set 3:

(1 comment)

> (1 comment)
 > 
 > So if the copying was broken, then how the test were passing? I
 > assume we had some tests testing this behavior; for example
 > instanceof should be broken due to missing type marker. I'm pretty
 > sure we had these tests so I think we broke the test itself in one
 > the refactoring.

I thought about this. Casting does not rely on copying j.l.Object properties. The castmap field is set by defineClass so Java subtypes of native classes have the right castmap even if object properties were not copied. The typeMarker is used for only a few things and the only consequence I think there is for casting is that these object will return true to instanceof JSO (will add in the test).

The only functionality that might break is toString, equals, hashcode where they would follow the JSO path instead. There is sure no coverage for hashcode and equals, and since in the end JSO and objects that don't override have basically the same behaviour most code would not notice the difference (of course we could construct a contrived test for this and we will)

Patch-set: 3
Label: Verified=0