Update patch set 6
Patch Set 6:

(4 comments)

Thanks for the followup and sorry for the… delay (sic) I believe one reason I didn't review it at the time was that I waited for someone (even) more versed in webappsec to chime in (and then I let it slip through the cracks). I'd still want that to happen btw.

I must say I, for one, am not a fan of the system property. I'd have rather kept the original idea of only adding 'tel:' support, and possibly add the configurability in a followup commit.

Patch-set: 6
Label: Verified=0