Update patch set 6
Patch Set 6:

(1 comment)

> (4 comments)
 > 
 > Thanks for the followup and sorry for the… delay (sic) I believe
 > one reason I didn't review it at the time was that I waited for
 > someone (even) more versed in webappsec to chime in (and then I let
 > it slip through the cracks). I'd still want that to happen btw.
 > 
 > I must say I, for one, am not a fan of the system property. I'd
 > have rather kept the original idea of only adding 'tel:' support,
 > and possibly add the configurability in a followup commit.

I tend to agree; a system property generally doesn't make for an API that's easy to reason about.


A bit of background on validating URLs, and safe schemes:

Indeed, tel: is safe to put on a whitelist with respect to XSS (a tel: URL won't result in script execution when dereferenced).

However, there is a potential concern with URL schemes that (esp on phones) result in actions that take the user out of the browser context into some other app.  This includes tel:, sms: android intents, play/app store schemes, and various application-specific schemes.  

Depending on the semantics of the web app in which an attacker-controlled link of such a scheme might occur, it could result in a situation that confuses the user, e.g. to place a call to a fraudulent premium-rate phone number ("1-900" number), or tricking them into installing a certain undesirable app.  In most cases this is unlikely to be a problem, but it does kind of depend on the app in question.

With that in mind, we've usually taken an approach of not blanket-whitelisting tel: and similar numbers in common validation APIs, but rather providing an API that allows a developer to specifically permit such schemes.  There is however a super-set whitelist that only contains schemes that will not execute script (i.e., there's no API that allows developers to allow javascript: and such).

For an example, see the implementation in the generalization of SafeHtml types, https://github.com/google/safe-html-types/blob/7adb4071373373772aa69db5a00d717bacaf5590/types/src/main/java/com/google/common/html/types/SafeUrls.java#L121 (*)

This approach seems preferable to adding entries to a whitelist via an ambient global flag, because it's easier to reason about during code review. 

FWIW, for schemes that occur rarely, it's usually preferable to write an application-specific SafeUrl factory method that does whatever custom validation or construction as necessary, and uses UriUtils.fromTrustedString to "bless" the result as SafeUri.  

 
(*) not sure if you've come across this implementation of SafeHtml types. They're basically an later re-implementation of the same concept, but independent of a particular UI toolkit or templating system.  It's a bit unfortunate that this sometimes results in apps that deal with both GWT and "general-purpose" SafeHtml types, but there really wasn't much of an alternative (it's not reasonable for non-GWT apps to depend on GWT just to get those types). In hindsight, we probably should have implemented SafeHtml in friends as a separate library to begin with rather than baking them into the c.g.gwt namespace, but alas, it's too late for that...


 > (4 comments)
 > 
 > Thanks for the followup and sorry for the… delay (sic) I believe
 > one reason I didn't review it at the time was that I waited for
 > someone (even) more versed in webappsec to chime in (and then I let
 > it slip through the cracks). I'd still want that to happen btw.
 > 
 > I must say I, for one, am not a fan of the system property. I'd
 > have rather kept the original idea of only adding 'tel:' support,
 > and possibly add the configurability in a followup commit.

Patch-set: 6