Update patch set 22
Patch Set 22:

(2 comments)

> > I still have the performance and code-size issue
 > 
 > The removement of the MAILTO case should catch it...

I meant that unless the GWT compiler can optimize them out, we're "forcing" everyone to loop over a (possibly empty) set, and creating that set (there might be cases where people have absolutely no call to Collections.emptySet()). I hope the GWT compiler is capable of removing all the extra code about the "extra allowed schemes" if one doesn't use them, but I'd like a factual confirmation (e.g. simple app compiled with 2.8.0 and with the patch applied, and comparing the output)

The gains with the removed MAILTO case are independent of that change (could be done in a separate commit that benefits everyone); they also only apply, in terms of potential runtime perfs, for URLs that would be unsafe, and the common case should be http and https (or scheme-less, relative) URLs, which would be unaffected.

Patch-set: 22
Label: Verified=0