Fix bad rewriting of default methods.
ReplaceDefenderMethodRefernces rewrites default methods into static
methods and rerouting all the explicit callsites. The callsite rewrite
was incorrectly assuming that calls would always have JThisRef as a
qualifier, which is true initially but not in the midst of the
rewriting process.

A second issue was the rewriting of callsites might have been
incomplete as new methods are created by this process and might or
might not be themselves processed to rewrite default method dispatch.

Bug: #9453
Bug-Link: https://github.com/gwtproject/gwt/issues/9453
Change-Id: I4ee3b340ce780cd0458920c9951b442fc25f1523