Update patch set 2
Patch Set 2:

Frustratingly, the more I dig, the worse this gets - this was originally reported in the gwtproject/gwt's gitter chat or IRC, and a bug was going to be filed for it (though I haven't seen it). I can confirm that for libraries that are unable to use Array.isArray (IE8 support for instance, which GWT ostensibly supports) without enumerating each frame and testing `object instanceof frame.Array`. Which this code doesn't do anyway, so JSONParser.java is incompatible with itself, running in two different GWT modules.

AutoBeans do this correctly, and elemental 1's JsJsonFactory does as well (though its Json class does not, surprisingly).

And since JavaScriptObject's various static create methods do not reference $wnd, no code that doesn't manually reimplement that JSNI will be wrong as well, even within these modified classes (see JSONArray's, JSONObject's no-arg ctors). 

With regard to existing code however, if someone is depending on only using the in-frame prototypes and constructors with JSOs anyway and are unaware of the need to test against the base window, I'd suggest we'd rather have code that is largely correct than surprisingly broken - in the same way that supporting Object[]'s ability to push a new value instead of creating a new array that is broken and only works by depending on undocumented details that should not be relied on to never change, or code which assumes that object.tS() will forever be toString().

So with regard to testing, I'm unconcerned unless we're prepared to make many more undocumented implementation details also hard and fast rules, unless there is a specific doubt I should be looking after in tests. However, with JavaScriptObject/JSONArray also being potentially being incompatible with itself, I'm going to do a little more digging before I expect I'll have to abandon this.

Or greatly expand its scope, which seems unlikely.

Patch-set: 2