Update patch set 2
Patch Set 2:

> I'm confused - the hack in Guava being done up in JSNI wouldn't help J2CL anyway, right? Or is JSNI (or some analog) supported in J2CL now in such a way that the Guava hack remaining is better than this for both GWT and J2CL?

Yes, J2CL has some kind of JSNI replacement for edge cases. It is supplied as an extra file so you don't see it open-source Guava.

> Also, even though this patch may not help Guava specifically, isn't it still an improvement for any GWT usage of Objects.equals taking string parameters?

Who said that? If you are referring to my comment, I only commented on an incorrect statement inside a commit msg.

Patch-set: 2