Update patch set 2
Patch Set 2:

Perhaps he assumed (as I did) that com.google.common.base.Objects was on its way out since the next Guava is said to be dropping support for older versions of Java, and both Objects.hashCode and equals are deprecated for use in Java 7 and above. 

At least that is what I assumed this was in reference to, since ExtraObjectsMethodsForWeb adds a specialized overload of equals just for String, which is what this class aimed to remove the need for.

Patch-set: 2