Update patch set 5
Patch Set 5:

> Without yet running the tests specifically, this may be a case
 > where com.google.gwt.core.client.GwtScriptOnly needs to be applied.
 > 
 > However, I had been under the impression that classic dev mode
 > skipped all supersource anyway, so we wouldn't need to specify that
 > our alternate implementations shouldn't be used. This may not bethe
 > case - for example, com.google.gwt.core.client.impl.WeakMapping's
 > supersource uses @GwtScriptOnly.

Actually I pointed this will not work in devmode earlier in thread.
Devmode is already not using this version of the class. This is about a call from outside being bound to a method that is not available in JRE. We should simply apply Donorunwith to test method.

Patch-set: 5