Fix bad varargs conversion in method reference.
The functional method in a functional interfaces and the method
reference implementing it do not need to be exact matches
parameterwise.

The compiler should insert conversion code to bridge that gap.

In particular take the code

  String first(Strings... strings) { return strings[0]; }

  Function<String, String> f = this::first;

  f("Hello");

The class to f in this case needs some conversion as the implementation
expectss a String...  (which is exacty a String[]).

The code that performed the conversion in GwtAstBuilder had a bug when
the functional type was generic.

Bug: #9497
Bug-Link: https://github.com/gwtproject/gwt/issues/9497
Change-Id: I3ea724a9586dd53e1e94591b5685475bb8000e40