Fixes #9504: Corrects Array.sort(double[]) and Array.sort(float[])
This adds a test for double and float sorting that includes NaNs
and infinite values, and corrects the implementation of Arrays.sort()
to properly handle floating point arrays consistent with the Java API
specification which states that NaNs are considered greater than all
other values for the purposes of sorting.

This patch does not address the ordering of -0 and +0 which I don't
see a way to reasonably implement in JavaScript.

Change-Id: I2ad9abffbd38e70ba7d5a20011392d879383b7b8