Make Array.sort(float[]/double[]) JDK compliant
This adds a test for double and float sorting that includes NaNs
and infinite values, and corrects the implementation of Arrays.sort()
to use Double.compare() and Float.compare(), which properly handle
NaN values.

This also corrects Arrays.equals(double[]/float[]) to treat NaN
values as equal for the purpose of the Arrays.equal() method.

Bug: #9504
Bug-Link: https://github.com/gwtproject/gwt/issues/9504
Change-Id: I2ad9abffbd38e70ba7d5a20011392d879383b7b8