Update patch set 1
Patch Set 1:

> What would it take to fix them rather than just disable them? How
 > many are there?

There are quite a few and they only matter when we are using our internal build since those warnings are considered errors. Those settings of course don't affect at all the opensource build process.

I am not sure that is worth cleaning the code to avoid the warnings, that said if someone want to run ErrorProne on the opensource build and fix the code, that would be fine.

Patch-set: 1
CC: Gerrit User 5305 <5305@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>