Update patch set 3
Patch Set 3:

Still couldn't find a nice workaround. I can change it to Void but that still allows you to call Object methods on it and pass around. Maybe that's enough though.

If we ever going to allow this, maybe we should at least allow it if the method name is debugger.

All aside I wonder if js optimization passes in GWT already drop this since it is simply a property access?

Patch-set: 3