Update patch set 2
Patch Set 2:

> Since we just cut 2.8.1 please leave trunk clean of any non fixes
 > for 2.8.1, so that we can still merge patches if necessary to get
 > 2.8.1 out.

Well, this code was already merged in master, and reverted about one hour ago because of check-style issues that break jenkins build, so I was reapplying after fixing those aesthetic issues.
This issue is marked as 2.8.1 milestone in github btw.
What is the last commit in the 2.8.1 cut, does it includes the original code of this?

Patch-set: 2
Reviewer: Gerrit User 5055 <5055@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>