Update patch set 3
Patch Set 3:

> https://groups.google.com/d/topic/google-web-toolkit-contributors/GIG-3-0Bp1E/discussion

Sure, I was already aware of this. I meant this info was generic, there was no info when the cut was going to happen

 > It had been reported first 3 years ago already, patch sent on
 > Gerrit 18 months ago, so I believe there's no hurry.
 > 
 > If you ask me, this is also a symptom of some bad coding/architecture
 > choice (no, really, a localizable map with 2000 entries? not
 > something we'd really want to support/encourage), and there are
 > workarounds if really needed.

I think the discussion is not whether we encourage or not certain practices.
I agree that the patch was queued for a very long, basically because the original reviewer probable din't take the time to continue with the discussion.
The fact is that the author asked lately for the patch status and updated it. Everything looks  ok, and there were no breaking changes, and no-one else voted against it.

Anyways I dont have any strong preference to merge this, but that it is an external contribution needed for their use cases.

Patch-set: 3
Reviewer: Gerrit User 5055 <5055@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>