Update patch set 3
Patch Set 3:

@Thomas
I used JUnit4 because its the latest stbel version and is my preference.

 > It had been reported first 3 years ago already, patch sent on  
 > Gerrit 18 months ago, so I believe there's no hurry.

We use a patched 2.7 version until yet but want to update to GWT 2.8 and it would be nice if this is included. If not we have to patch it again

 > > could you maybe set the author to the actual author of the patch?
 > 
 > Changed the author of the commit
 > 
 > > Why haven't the tests been written in JUnit 3? I don't see any
 > > reason to use JUnit 4 here.
 > 
 > It was the author preference, since GWT build already depends on 
 > jUnit-4 I didn't see any strong need to use JUnit3 syntax, or, do
 > we have some rules/conventions to prevent using JUnit4 in our
 > tests?

 >  a localizable map with 2000 entries?
We use a properties file with many properties. Sure we can refactor this file but the method should either compile without warning or forbid this situation and fail the compilation. I think the working method is the best approcach

Patch-set: 3
Reviewer: Gerrit User 6910 <6910@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>