The onload handler is used to detect the situation when an image has loaded before having been "attached" (added to the page). This is not really possible to do another way. Although the DOM APIs define a 'complete' property which can be used to determine if the image has finished loading, there is no similar property that will tell you if an error occurred along the way.
This CL removes the onload property from the SafeHtmlTemplate. This
should be safe to do because the template is supposed to be treated
opaquely (according to AbstractImagePrototype.getSafeHtml() docs) and
therefore should not have any listeners attached.

The .createStructure() method is refactored to use DOM APIs to create
the image element and set the onload property in a CSP-compatible way.

Change-Id: Iefdf49204b83ef836d37492144b27e636d513f17