Forward window.onerror into the UncaughtExceptionHandler.
This change allows GWT to trap window.onerror and forward
errors to the uncaught exception handler. This is necessary
for libraries like elemental2 so that users do not constantly
have to call the entry function manually.
Right now this is still disabled and a subsequent change will
enable it by default.

For this change there are three modes:
 - Trap window on error at all times and forward
 - Only trap window.onerror if there is no current handler
 - Do not trap anything (old behavior)

With this change some applications will see more calls
to their UncaughtExceptionHandler since these would get
ignored beforehand, but we strongly suggest that this is
still the right way forward and users should fix those errors.
For the transition period we added the above switches.

Change-Id: I54f979c1887f44791f5bd2f32190b692f421fcce
Review-Link: https://gwt-review.googlesource.com/#/c/18540/