Update patch set 5
Patch Set 5:

It occurs to me that there might be a race condition with how we are setting the onload handler. Depending on how fast the image loads and implementation details, it could already be done loading by the time the 'onload' handler is set, in which case it would not fire.

It would be safer to set the 'onload' property first and only then set the 'src' property. Would be okay to revert to the programmatic style of the previous iteration to implement this?

Patch-set: 5
Reviewer: Gerrit User 10840 <10840@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>