Update patch set 2
Patch Set 2:

I agree with Thomas, we should not pro-actively add a ClickHandler to make Anchor CSP compatible. 

As already already mentioned, if you use an Anchor but never set a real HREF on it then you should use a button because you are dealing with an UI action and not with an URL that can be  opened in a new browser window. 

I would love to see a contribution that deprecates the use of javascript:; href inside Anchor (deprecate constructor, log warning for setHref("javascript:;") during development) and provides an ActionLink widget that looks like an Anchor but is based on Button.

Patch-set: 2
Reviewer: Gerrit User 5031 <5031@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>