Update patch set 2
Patch Set 2:

The problem I'm dealing with is adding CSP to existing, legacy apps. I would like to minimize my effort here and not have to refactor Anchors to Buttons, etc. The goal here is to be able to enable CSP on an existing GWT apps and have things just work. That would be ideal and would save tons of time for us security folks and developers that wish to adopt CSP for their product.

The suggestions to deprecate and tell developers to buttons, etc are fine but can we please not gate CSP adoption on them. If there is any way that we can preserve existing behavior while being CSP compatible, that is the way to go.

Patch-set: 2
Tag: mailMessageId=<CAJGyu7VnMSNhxPpnPD4tQsGOTYTLa9URsTGOuBz6YN8KEBdF9w@mail.gmail.com>