Update patch set 3
Patch Set 3:

It would be possible to condition the handler on an opt-in configuration property. It would be a little bit tricky to configure for some projects. For example, Apps Script has a 8 different .gwt.xml files and it's not obvious which one to change.

Is the concern that the handler could lead to breakage? If so, it would still be bad if something broke for an opt-in property and we should make sure the handler is safe. Is the concern perf? Maybe we could logic to add the handler only when it is necessary. I would much prefer we find a solution could be on by default and would be safe. It would also be awesome to have GWT be CSP compatible out of the box.

Patch-set: 3
Tag: mailMessageId=<CAJGyu7Vo8wi_it5yQjYzaA5fRpvXHZ_0HBgwoE0VUqw_a1-MVQ@mail.gmail.com>