Replace String.equals with == in InternalPreconditions.
Due to cycle between java.lang.String and InternalPreconditions
it requires more parsses to turn precidition flags into static
constants.
This also prevents equals to be fully optimized as it results in a
checken-egg problem in J2CL (equals needs checkNotNull to be removed
to be evaluated but checkNotNull cannot be removed before the equals
calls in preconditions evaluated).
GWT doesn't hit this problems because String.equals is special cased
for optimization purposes.

Change-Id: Ib32bb141f98477cde7b8d56689a86829e1cf0a94