Add support for detecting CSP violations during unit tests in GWT.
This CL defines and implements a config property
"strictCspTestingEnabled" that can be set in .gwt.xml
config files. When running unit tests using a module where
this option is enabled, the GWT test runner enforce a strict
(https://csp.withgoogle.com/docs/strict-csp.html) CSP policy on the
running page and will detect and report any violations and the test
the caused them.

In the past, we've made fixes to GWT to make it CSP-compatible but have
not had a good way to test those fixes. This CL provides effective way
to ensure those fixes are correct and will not be accidentally undone
or broken by future changes.  By enabling this option, we can also
pro-actively detect CSP violations in the codebase.  This option can
also be used by GWT users to test their CSP fixes and ensure their
codebase is CSP-compatible.

A report-only CSP header is added to junit.html pages in
JUnitShell.java. This is only added to modules that have the cspTesting
option enabled. A handler to catch the resulting CSP violations is
added to JUnitHostImpl.java.

Because CSP violations are generated and sent asynchronously, it is
hard to attribute a violation to the test which generated it. By the
time we get a violation, multiple tests may have already started and
finished. It is also hard to know for how long to wait for straggler
violations to come in after a test is finished.

To solve this, we generate a "checkpoint" CSP violation with
details about the just-completed test after every test execution.
CSP violations that are received are put in a pending list (in
JUnitMessageQueue.java).  Once a checkpoint violation is received,
the pending violations are saved in the appropriate result object
and the pending violations list is cleared.

Change-Id: I387de41fb2019086c741d9c81077f54b0e8595ea